% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimura-distribution.R
\name{.f0}
\alias{.f0}
\title{Probability of allele loss}
\usage{
.f0(p, b)
}
\arguments{
\item{p}{number in (0; 1): initial heteroplasmy}

\item{b}{number in (0; 1): time / rate parameter}
}
\value{
The probability the allele is lost due to genetic drift.
}
\description{
Estimates the probability that an allele is lost due to random genetic drift.
}
\examples{
.f0(0.3, 0.5)
}
