% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimura-distribution.R
\name{.f1}
\alias{.f1}
\title{Probability of fixing allele}
\usage{
.f1(p, b)
}
\arguments{
\item{p}{number in (0; 1): initial heteroplasmy}

\item{b}{number in (0; 1): time / rate parameter}
}
\value{
The probability the allele is fixed due to genetic drift.
}
\description{
Estimates the probability that an allele is fixed (i.e. the wildtype is lost)
due to random genetic drift.
}
\examples{
.f1(0.3, 0.5)
}
