% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift.R
\name{.get_ks_stat}
\alias{.get_ks_stat}
\title{Get the KS test statistic between a heteroplasmy CDF and a Kimura CDF}
\usage{
.get_ks_stat(h, cdf_lt)
}
\arguments{
\item{h}{numeric: heteroplasmy fractions}

\item{cdf_lt}{list of length 2, see pkimura_full}
}
\value{
Positive D
}
\description{
Get the KS test statistic between a heteroplasmy CDF and a Kimura CDF
}
\examples{
h <- seq(0.2, 0.5, 0.01)
cdf_lt <- fit_kimura(h)$cdf
.get_ks_stat(h, cdf_lt)
}
