% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimura-distribution.R
\name{.phi}
\alias{.phi}
\title{Density of the Kimura distribution}
\usage{
.phi(x, p, b)
}
\arguments{
\item{x}{number in (0; 1): the point at which F() is calculated.}

\item{p}{number in (0; 1): initial heteroplasmy}

\item{b}{number in (0; 1): time / rate parameter}
}
\value{
The density of Kimura(p, d), evaluated at x.
}
\description{
Estimates the probability density function at x for heteroplasmy under no
selection pressure.
}
\examples{
.phi(0.2, 0.3, 0.5)
}
