% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-pars.R
\name{.pkimura_full}
\alias{.pkimura_full}
\title{The full Kimura distribution CDF}
\usage{
.pkimura_full(p, b)
}
\arguments{
\item{p}{number in \eqn{(0; 1)}: initial heteroplasmy.}

\item{b}{number in \eqn{(0; 1)}: drift parameter.}
}
\value{
Numeric vector of length 10001, corresponding to the CDF for
\code{Kimura(p, b)}, approximated at \code{x = seq(0, 1, 1e-4)}.
}
\description{
Get the CDF for \code{Kimura(p, b)} evaluated at regular intervals of length
1e-04. Note: this is an auxiliary function.
}
\examples{
.pkimura_full(0.3, 0.5)
}
\references{
Wonnapinij, Passorn, Patrick F. Chinnery, and David C. Samuels.
"The distribution of mitochondrial DNA heteroplasmy due to random genetic
drift." The American Journal of Human Genetics 83.5 (2008): 582-593.
}
