% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift.R
\name{.sample_h}
\alias{.sample_h}
\title{Sample heteroplasmies from a given Kimura CDF}
\usage{
.sample_h(n, cdf_lt, round = TRUE)
}
\arguments{
\item{n}{number of observations to generate.}

\item{cdf_lt}{list of length 2, see pkimura_full}

\item{round}{logical; whether to round to 2 decimal places}
}
\value{
vector length n
}
\description{
Sample heteroplasmies from a given Kimura CDF
}
\examples{
h <- seq(0.2, 0.5, 0.01)
cdf_lt <- fit_kimura(h)
.sample_h(length(h), cdf_lt)
}
