% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-pars.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{Estimate Kimura distribution parameters}
\usage{
estimate_parameters(h)
}
\arguments{
\item{h}{numeric vector in \eqn{[0; 1]} of heteroplasmy fractions.}
}
\value{
A named vector of length two with elements \code{p} and \code{b}.
These are the estimated parameters of a \eqn{Kimura(p, b)} distribution to
fit to the data.
}
\description{
Given a vector of heteroplasmy fractions, estimate the Kimura distribution
parameters.
}
\examples{
h = stats::runif(10)
estimate_parameters(h)
}
