% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift.R
\name{fit_kimura}
\alias{fit_kimura}
\title{Fit a Kimura distirbution to heteroplasmy data}
\usage{
fit_kimura(h)
}
\arguments{
\item{h}{numeric: heteroplasmy fractions}
}
\value{
List of length 3 = p, b, and the estomate
}
\description{
Fit a Kimura distirbution to heteroplasmy data
}
\examples{
h <- seq(0.2, 0.5, 0.01)
fit_kimura(h)
}
