% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{nvar}
\alias{nvar}
\title{Normalised heteroplasmy variance}
\usage{
nvar(H, percent = TRUE)
}
\arguments{
\item{H}{numeric vector of heteroplasmy fractions, either in \eqn{[0; 1]} or
represented as percentages in \eqn{[0; 100]}}

\item{percent}{logical; are values in H represented as percentage points?}
}
\value{
The normalised heteroplasmy variance, which is typically in
\eqn{[0; 1]}. Lower values imply lower variance and less genetic
drift, with 0 corresponding to all values of H being equal.

Note \code{nvar(H)} is equivalent to \code{1-estimate_parameters(H)["b"]}.
}
\description{
Normalised heteroplasmy variance
}
\examples{
# Using percentage points
H <- sample(0:100, 10)
nvar(H)

# Using heteroplasmy fractions between 0 and 1
h <- stats::runif(10)
nvar(h, percent = FALSE)
}
