% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimura-distribution.R
\name{pkimura}
\alias{pkimura}
\title{The Kimura distribution}
\usage{
pkimura(x, p, b)
}
\arguments{
\item{x}{number in \eqn{(0; 1)}: the point at which the density is
calculated.}

\item{p}{number in \eqn{(0; 1)}: initial heteroplasmy.}

\item{b}{number in \eqn{(0; 1)}: drift parameter.}
}
\value{
The CDF of Kimura(p, d), evaluated at x.
}
\description{
Estimates the Kimura cumulative distribution function.
}
\examples{
pkimura(0, 0.3, 0.5)
pkimura(0.2, 0.3, 0.5)
pkimura(1, 0.3, 0.5)
}
