% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-pars.R
\name{pkimura_full}
\alias{pkimura_full}
\title{The Kimura distribution}
\usage{
pkimura_full(p, b)
}
\arguments{
\item{p}{number in \eqn{(0; 1)}: initial heteroplasmy.}

\item{b}{number in \eqn{(0; 1)}: drift parameter.}
}
\value{
list of length 2, x and y
}
\description{
Get complete CDF for Kimura(p, b).
}
\examples{
pkimura_full(0.3, 0.5)
}
