% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{shift}
\alias{shift}
\title{Normalised heteroplasmy shift}
\usage{
shift(H, H0, percent = TRUE, tol = 1e-04)
}
\arguments{
\item{H}{numeric vector of heteroplasmy fractions, either in \eqn{[0; 1]} or
represented as percentages in \eqn{[0; 100]}}

\item{H0}{numeric vector of baseline heteroplasmy fractions, either in
\eqn{[0; 1]} or represented as percentages in \eqn{[0; 100]}}

\item{percent}{logical; are values in H represented as percentage points?}

\item{tol}{tolerance with which to resolve homoplasmy}
}
\value{
A vector of normalised heteroplasmy shift calculated as

\eqn{shift(H; H_0) = \frac{H(1-H_0)}{H_0(1-H)}}.

Homoplasmic states are resolved as \code{tol} for homoplasmic wildtype, and
\code{(1 - tol)} for homoplasmic mutant.
}
\description{
Normalised heteroplasmy shift
}
\examples{
# Using percentage points
H <- sample(0:100, 10)
H0 <- 50
shift(H, H0)

# Using fractions
h <- stats::runif(10)
h0 <- 0.3
shift(h, h0, percent = FALSE)
}
