% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift.R
\name{test_kimura}
\alias{test_kimura}
\title{Monte Carlo KS test for genetic drift}
\usage{
test_kimura(h, num_MC = 1000, round = TRUE)
}
\arguments{
\item{h}{heteroplasmy}

\item{num_MC}{number MC simulations}

\item{round}{whether the het is rounded}
}
\value{
object of class htest
}
\description{
Monte Carlo KS test for genetic drift
}
\examples{
h <- seq(0.2, 0.5, 0.01)
test_kimura(h)
}
