% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimura-distribution.R
\name{dkimura}
\alias{dkimura}
\title{The Kimura distribution}
\usage{
dkimura(x, p, b)
}
\arguments{
\item{x}{number in (0; 1): the point at which F() is calculated.}

\item{p}{number in (0; 1): initial heteroplasmy}

\item{b}{number in (0; 1): time / rate parameter}
}
\value{
The density of Kimura(p, d), evaluated at x. Note there is a fixed
probability at both extremes of the distribution.
}
\description{
Estimates the Kimura probability density function.
}
\examples{
dkimura(0, 0.3, 0.5)
dkimura(0.2, 0.3, 0.5)
dkimura(1, 0.3, 0.5)
}
