% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimura-distribution.R
\name{.hypgeo}
\alias{.hypgeo}
\title{Estimate F(i-1, i+2, 2, x)}
\usage{
.hypgeo(i, x)
}
\arguments{
\item{i}{positive integer: the index in the infinite sum of the Kimura distribution.}

\item{x}{number in (0; 1): the point at which F() is calculated.}
}
\value{
The value of F(i-1, i+2, 2, x), estimated recursively.
}
\description{
Estimates the hypergeometric function F(i-1, i+2, 2, x) needed for
calculating the Kimura distribution.
}
\examples{
.hypgeo(5, .3)
}
