% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-pars.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{Estimate Kimura parameters}
\usage{
estimate_parameters(h)
}
\arguments{
\item{h}{numeric: heteroplasmy fractions}
}
\value{
A named vector of length two.
}
\description{
Given a vector of heteroplasmies, estimated the Kimura distribution
parameters.
}
\examples{
h = stats::runif(10)
estimate_parameters(h)
}
