% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kimura-distribution.R
\name{rkimura}
\alias{rkimura}
\title{The Kimura distribution}
\usage{
rkimura(n, p, b)
}
\arguments{
\item{n}{number of observations to generate.}

\item{p}{number in (0; 1): initial heteroplasmy}

\item{b}{number in (0; 1): time / rate parameter}
}
\value{
A vector of length n.
}
\description{
Generates from the Kimura distribution.
}
\examples{
rkimura(10, 0.3, 0.5)
}
